+++ tests/collection_tests/test_dicts.py
@@ -1,5 +1,6 @@
 import pytest
 from rics.collections.dicts import (
+    DictComparer,
     InheritedKeysDict,
     compute_if_absent,
     flatten_dict,
@@ -136,7 +137,7 @@ def flattened_dict():
 
 
 @pytest.fixture
-def d():
+def ikd():
     return InheritedKeysDict.make(
         dict(
             default={0: "shared0", 1: "shared1"},
@@ -148,17 +149,65 @@ def d():
     )
 
 
-def test_inherited_keys_dict(d):
-    assert d["ctx0"] == {0: "c0-v0", 1: "shared1"}
-    assert d["ctx1"] == {0: "c1-v0", 1: "c1-v1"}
-    assert d["not-in-d"] == {0: "shared0", 1: "shared1"}
-    assert len(d) == 2
-    assert list(d) == ["ctx0", "ctx1"]
+def test_inherited_keys_dict(ikd):
+    assert ikd["ctx0"] == {0: "c0-v0", 1: "shared1"}
+    assert ikd["ctx1"] == {0: "c1-v0", 1: "c1-v1"}
+    assert ikd["not-in-d"] == {0: "shared0", 1: "shared1"}
+    assert len(ikd) == 2
+    assert list(ikd) == ["ctx0", "ctx1"]
 
     with pytest.raises(KeyError):
         empty: InheritedKeysDict[str, str, str] = InheritedKeysDict()
         _ = empty["any-key-will-do"]
 
 
-def test_copy(d):
-    assert d == d.copy()
+def test_copy_inherited_keys_dict(ikd):
+    assert ikd == ikd.copy()
+
+
+class TestDictComparer:
+    def test_added_only(self):
+        dc = DictComparer({-1: None, 1: 1}, baseline={i: i for i in range(3)})
+        assert dc.total == 3
+        assert dc.changed == {-1: None, 0: 0, 2: 2}
+        assert dc.added == {-1: None}
+
+    @pytest.mark.parametrize("preference", ["old", "new"])
+    def test_updated_only(self, preference):
+        if preference == "new":
+            expected_zero_value = None
+        else:
+            expected_zero_value = 0
+        expected_str = f"updated(1): {{0: {expected_zero_value}}}, removed(1): {{2}}"
+
+        dc = DictComparer({0: None, 1: 1}, baseline={i: i for i in range(3)}, preference=preference)
+        assert dc.preference == preference
+        assert dc.total == 2
+        assert dc.changed == {0: expected_zero_value, 2: 2}
+        assert dc.updated == {0: expected_zero_value}
+        assert dc.to_string() == expected_str
+
+    def test_removed_only(self):
+        dc = DictComparer({0: None, 1: 1}, baseline={i: i for i in range(3)})
+        assert dc.total == 2
+        assert dc.changed == {0: None, 2: 2}
+        assert dc.removed == {2: 2}
+        assert dc.to_string() == "updated(1): {0: None}, removed(1): {2}"
+
+    def test_all(self):
+        dc = DictComparer({0: None, 1: 1, -1: None}, baseline={i: i for i in range(3)})
+        assert dc.preference == "new"
+
+        assert dc.total == 3
+        assert len(dc.changed) == dc.total
+        assert len(dc.changed) == len(dc.added) + len(dc.updated) + len(dc.removed)
+
+        assert dc.added == {-1: None}, dc.added
+        assert dc.updated == {0: None}, dc.updated
+        assert dc.removed == {2: 2}, dc.removed
+        assert dc.changed == {-1: None, 0: None, 2: 2}
+        assert dc.to_string() == "added(1): {-1: None}, updated(1): {0: None}, removed(1): {2}"
+
+    def test_wrapper(self):
+        actual = DictComparer({"foo": {"bar": 1}, "baz": 0}, baseline={"foo": 1, "bar": None}).to_string(flatten=True)
+        assert actual == "added(1): {'baz': 0}, updated(1): {'foo.bar': 1}, removed(1): {'bar'}"