+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * New module `rics.paths`; `paths.parse_any_path()`, derived functions `any_path_to_str()` and `any_path_to_path()`.
 * New module `rics.strings`. Moved `format_perf_counter()` and `format_seconds()`.
 * New function `strings.format_bytes()`.
+* New function `rics.collections.dicts.format_changed_keys()`.
 
 ### Deprecated
 * Module `rics.ml.time_split` use `time-split` [![PyPI - Version](https://img.shields.io/pypi/v/time-split.svg)](https://pypi.python.org/pypi/time-split) package instead.